==> 01-Build-Fig-Code.Rmd <==
---
title: "01-Explore-Data"
author: "Kurt Ingeman"
date: "4/7/2022"
output: html_document
---

## Load Libraries

```{r}
library(here)
library(tidyverse)
library(ggplot2)
library(rlang)
library(hrbrthemes)

```

# Clear Environment

```{r}
rm(list = ls())
```


# Read data (template)

```{r}
df.csv <- read.csv(here("Data", "Raw", "Juris", "juris.estimates.csv"))
df.depth.csv <- read.csv(here("Data", "Raw", "Juris", "juris.depth.estimates.csv"))

df.rds <- readRDS("~/github/JCR-FMP/Data/Raw/Strata/priorityspp.size.meancount.site.strataN2.rds") %>% 
  filter(mean.count > 0) %>% 
  mutate(TL = as.numeric(Size_cmTL)) %>% 
  mutate_if(is.character, as.factor)


df.rds$TAXONNAME <- as.factor(recode(df.rds$TAXONNAME, "Chlorurus spilurus/sordidus" = "Chlorurus spilurus")) 

df.rds$ISLAND <- droplevels(df.rds$ISLAND)
  
```


# Explore Site-level data across dataset (*WARNING*)

## *WARNING* For exploration purposes only -- data sampled unevenly across strata so summing to larger scales than stratum could lead to spurious results (if size is systemattically biased by depth, location, or other) 

```{r}

length(df.rds$mean.count) # 2,598,343 obs including zeroes # 21,662 excluding 0 
length(unique(df.rds$TAXONNAME)) # 46 priority species
```

# Make a tidy dataframe from the juridictional summary df

```{r}

df.tidy <- df.csv %>% 
  filter(Fish.grp == "priority" & pool.mean > 0)

str(df.tidy)

```


# Write plot functions in two ways 
```{r}
# Curly curly
scatter_plot <- function(data, x, y, ...) {
  ggplot(data, aes({{x}}, {{y}})) +
    geom_point()
}
scatter_plot(df.tidy, pool.mean, pool.var)

# Taking data and mapping as the functions arguments 
scat_map <- function(data, mapping) {
  ggplot(data, mapping) +
    geom_point(alpha = 0.5)
}

scat_map(df.tidy, aes(pool.mean, pool.var, size = pool.var, color = REGION))
# Hey

```

# Identify  couple of values with outlier variance

```{r}
high.var <- df.tidy %>% 
  filter(pool.var > 0.00002) # a couple of ACGU size classes from 2018
```

# Play with the raw count data 

## Basic density plots and histograms

```{r}
str(df.rds)

# basic density plot and histogram
ggplot(df.rds, aes(x=TL)) + 
  geom_density(alpha = 0.4) 

ggplot(df.rds, aes(x=TL)) + 
  geom_histogram(alpha = 0.4)

# can't overlay them because the counts are huge relative to frequency

# a prettier
ggplot(df.rds, aes(x=TL)) + 
  geom_histogram(binwidth=1, 
                 fill="#002B7F", 
                 color="#e9ecef", 
                 alpha=0.7) +
  ggtitle("Bin size = 1") +
  theme_ipsum() +
  theme(
  plot.title = element_text(size=15)
    )

```

# Make a function for density plot and histogram
```{r}

density_plot <- function(data, x, alpha) {
  ggplot(data, aes({{x}})) +
    geom_density(fill = "#C71B36", 
                 alpha = alpha) +
  theme_ipsum()
}

density_plot(df.rds, TL, 0.6)

# Here, I specific the fill and title outside the fucntion
hist_plot <- function(data, x, bin, fill, title) {
  ggplot(data, aes({{x}})) +
   geom_histogram(binwidth = bin, 
                 fill = fill, 
                 color="#e9ecef", 
                 alpha=0.7) +
  theme_ipsum() +
  ggtitle(title) +
  theme(
  plot.title = element_text(size=15))
}

my.fill = "#008DE8"

hist_plot(df.rds, TL, 2, my.fill, "Size Frequency: All species")


```

####################################################################################################################################################### Turn counts within a size at a particular site in obs #######
##############################################################################################################################################

```{r}

# dummy data set, turn a count variable into rows with uncount()
df <- tibble(
  meta = c("here", "here", "everywhere"),
  size = c(10, 11, 12),
  mean.count = c(3, 1, 2))

df %>% uncount(mean.count) %>% 
  as_tibble

# beautiful, now I just have the problem of fractional observations due to observations in one cylinder

# try multiplying all counts by factor that produces an integer (would just need to change axes by the same factor)
df <- data.frame(
  meta = c("here", "there", "everywhere"),
  size = c(10, 11, 12),
  mean.count = c(2, 1, 0.5)) %>% 
  mutate(mean.count = 2*mean.count) %>% 
  uncount(mean.count) 

# IMPORTANT this is fine for visualization purposes (provided frequency or rescale counts) but not fine for applying any statitical test as I have inflated the number of observations by 

hist_plot(df, size, bin = 1, fill = my.fill, title = "Ignore count or\njust divide by 2")

# Figure out multiplier for fractional counts
unique(df.rds$mean.count[df.rds$mean.count < 1]) *4 
# the lowest possible multiplier is 4

```

# Apply uncount and multiplier to site-level data to produce (index of) individual fish observations
```{r}

df.counts <- df.rds %>% 
  mutate(index.count = 4* mean.count) %>% 
  select(-mean.count) %>% 
  uncount(index.count) 

```


# Generate means and plot overlayed 
```{r}
df.counts <- df.counts %>% 
  group_by(REGION) %>% 
  mutate(mu_reg = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% 
  group_by(ISLAND) %>% 
  mutate(mu_isl = mean(TL, na.rm=TRUE)) %>%
  ungroup() 

ggplot(data=df.counts, aes(x=TL, group=REGION, fill=REGION)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_reg, color = REGION),
             linetype = "dashed") +
      theme_ipsum()

ggplot(data=df.counts, aes(x=TL, group=ISLAND, fill=ISLAND)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_isl, color = ISLAND),
             linetype = "dashed") +
      theme_ipsum() +
   facet_grid(REGION ~ .)
  
# lots of islands and really similar distributions of size
# going to need a good way to send a list of islands through plot function

```

# Facet by region (island) instead 
```{r}
ggplot(data=df.counts, aes(x=TL, group=REGION, fill=REGION)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_isl, color = REGION),
             linetype = "dashed") +
  xlim(0,60) +
#      theme_ipsum() +
  facet_wrap(ISLAND ~ .)

ggplot(data=df.counts, aes(x=TL, group=TAXONNAME, fill=REGION)) +
    geom_density(adjust=1.5, alpha=.4) +
  xlim(0,60) +
#      theme_ipsum() +
  facet_wrap(TAXONNAME ~ .)


ggplot(data=df.counts, aes(x=TL, group=REGION, fill=REGION)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_reg, color = REGION),
             linetype = "dashed") +
 #    theme_ipsum() +
  facet_grid(REGION ~ .)


```
# Plot separate ggplot figures in a loop.

```{r}

# Make list of island names to loop over.

df_list <- split(df.counts, df.counts$ISLAND)

# Make individual density plots for each island
plot_list = list()

for (i in 1:length(df_list)) {
    p = density_plot(df_list[[i]], TL, 0.6)
    plot_list[[i]] = p
}


# Save plots to tiff. Makes a separate file for each plot.
for (i in 1:44) {
    file_name = paste("Figures/Size dist. ", names(df_list)[i], ".tiff", sep="")
    tiff(file_name)
    print(plot_list[[i]])
    dev.off()
}

# Another option is to use group_split and map() this time for each species
df.counts %>%
  mutate(OBS_YEAR = as.factor(OBS_YEAR)) %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, group = OBS_YEAR, fill = REGION) +
        geom_density(adjust=1.5, alpha=.4) +
        labs(x = "TL",
             y = "Frequency",
             title = first(.x$TAXONNAME)) +
       facet_grid(OBS_YEAR ~ REGION) +
        theme_classic() -> plot
      ggsave(paste0(first(.x$TAXONNAME), '_plot.png'), plot)
      })


# Next up: save into separate folders

```


# Simple  for-loop to create directories by species and fill with plots
```{r echo=FALSE}

df.spp <- split(df.counts, df.counts$TAXONNAME)

figure.dir <- "Figures/Scratch/"
project.dir <- "/Users/kurtingeman/github/JCR-FMP/"

for (i in 1:length(df.spp)) {
  folder <- paste0(figure.dir, names(df.spp[i]))
  if (!dir.exists(folder)) {dir.create(folder)}
  setwd(folder)
  p = density_plot(df.spp[[i]], TL, 0.6) # can add other plot from functions while we are here
  ggsave(filename = paste("Density", names(df.spp)[i], ".jpg", sep=" "),
         plot = p)
  q = hist_plot(df.spp[[i]], TL, 2, my.fill, paste("Size Frequency", names(df.spp)[i]))
  ggsave(filename = paste("Histogram", names(df.spp)[i], ".jpg", sep=" "),
         plot = q)
  setwd(project.dir)
}

# Huzzah

```

#######
# Build up from basic bar plots 
######


# Read in data
```{r}

df <- read.csv(here("Data", "Raw", "Juris", "juris.estimates.csv")) %>% 
  filter(Fish.grp == "priority" & pool.mean > 0) %>% 
  filter(Size.bin == "5 cmTL") %>% 
  select(!Fxn.grp) %>% 
  mutate(Size = factor(Size, levels = c("[0,5]", "(5,10]", "(10,15]", "(15,20]", "(20,25]", "(25,30]", "(30,35]", "(35,40]", "(40,45]", "(45,50]", "(50,55]", "(55,60]", "(60,65]", "(65,70]", "(70,75]", "(75,80]", "(80,85]", "(85,90]", "(90,95]", "(95,100]", "(105,110]", "(120,125]", "(125,130]","(135,140]"))) 

levels(df$Size)

```

# Build from simple barplot to clustered barplot

```{r}
df0 <- df %>% 
  filter(TAXONNAME == "Acanthurus lineatus") %>% 
  mutate(Year = as.factor(OBS_YEAR))


df1 <- df %>% 
#  filter(TAXONNAME == "Acanthurus lineatus") %>% 
  mutate(Year = as.factor(OBS_YEAR)) %>% 
  filter(REGION == "MARIAN")

df2 <- df1 %>% 
  filter(TAXONNAME == "Acanthurus lineatus")
  
  
ggplot(df0, aes(fill = as.character(Year), y=pool.mean, x=Size)) +
    geom_bar(position="dodge", stat="identity") +
    geom_errorbar(aes(x=Size, ymin=pool.mean - pool.se, ymax = pool.mean + pool.se), position="dodge", stat="identity") +
  facet_wrap(~REGION, scales = "fixed")

## need to loop through species for indiv plots on pages of pdf

```
# Create basic point and error plots 

## facet or looop through species and location

```{r}

ggplot(df0, aes(x=OBS_YEAR, y=pool.mean, fill = REGION)) +
    geom_line(color="grey") +
    geom_point(shape=21, color="black", size=3) +
   geom_linerange(aes(ymin=pool.mean - pool.se, ymax = pool.mean + pool.se), alpha=0.6, size=.6, position="dodge", stat="identity") +
    ggtitle("Mean biomass through time, by size bin") +
 facet_grid(Size ~ ., scales = "fixed")

```

# Create connected line plot for mean TL

## facet or looop through species and location

```{r}

str(df.counts)

df.counts <- df.counts  %>% 
  group_by(TAXONNAME, ISLAND, OBS_YEAR, DEPTH_BIN) %>% 
  mutate(mu_spp_isl_year_depth = round(mean(TL), 1)) %>% 
  ungroup() %>% ungroup() %>% ungroup() %>% ungroup()

```


```{r}
temp <- df.counts %>% 
  filter(TAXONNAME == "Acanthurus lineatus")

ggplot(temp, aes(x = OBS_YEAR, y = mu_spp_isl_year_depth, shape = REGION, color =  DEPTH_BIN, group = DEPTH_BIN)) +
    geom_line(color="grey") +
    geom_point(size=3, alpha = 0.5) +
  facet_wrap(~ISLAND)

## Loop through spp on different files or pdf


```
