==> 02-Size-Structure.Rmd <==
---
title: "02-Size-Structure"
author: "Kurt Ingeman"
date: "5/10/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load Libraries

```{r}
library(here)
library(tidyverse)
library(ggplot2)
library(rlang)
library(hrbrthemes)
library(scales)
library(gridExtra)
library(cowplot)
library(viridis)

```

# Clear Environment

```{r}
rm(list = ls())
```

# Define palettes and directories
```{r}
# guam (blue, red, green, yellow, light blue)
proj.pal <- c("#0072B2","#F0E442","#CC79A7", "#009E73",  "#E69F00", "#D55E00", "#999999")

proj.dir <- "/Users/kurtingeman/github/JCR-FMP/"

```


# Read data for territorial-scale size structure through time (not segregated by depth)

```{r}
jur <- read.csv(here("Data", "Raw", "Juris", "juris.estimates.csv")) %>% 
  filter(Fish.grp == "priority" & pool.mean > 0) %>% 
  select(!Fxn.grp) %>% 
  mutate(Year = as.factor(OBS_YEAR))

jur$TAXONNAME <- as.factor(recode(jur$TAXONNAME, "Chlorurus spilurus/sordidus" = "Chlorurus spilurus")) 

# split by bin width and relevel
jur5 <- jur %>% filter(Size.bin == "5 cmTL") %>% 
 mutate(Size = factor(Size, levels = c("[0,5]", "(5,10]", "(10,15]", "(15,20]", "(20,25]", "(25,30]", "(30,35]", "(35,40]", "(40,45]", "(45,50]", "(50,55]", "(55,60]", "(60,65]", "(65,70]", "(70,75]", "(75,80]", "(80,85]", "(85,90]", "(90,95]", "(95,100]", "(105,110]", "(120,125]", "(125,130]","(135,140]")))

jur10 <- jur %>% filter(Size.bin == "10 cmTL") %>%  
mutate(Size = factor(Size, levels = c("[0,10]","(10,20]", "(20,30]", "(30,40]", "(40,50]", "(50,60]", "(60,70]", "(70,80]", 
                                      "(80,90]", "(90,100]", "(100,110]", "(120,130]", "(130,140]")))

# Split by region and remove species from other region
mar5 <- jur5 %>% filter(REGION == "MARIAN")
mar5$TAXONNAME <-  droplevels(mar5$TAXONNAME)

mar10 <- jur10 %>% filter(REGION == "MARIAN")
mar10$TAXONNAME <-  droplevels(mar10$TAXONNAME)

sam5 <- jur5 %>% filter(REGION == "SAMOA")
sam5$TAXONNAME <-  droplevels(sam5$TAXONNAME)

sam10<- jur10 %>% filter(REGION == "SAMOA")
sam10$TAXONNAME <-  droplevels(sam10$TAXONNAME)

```


# Function for clustered barplot of biomass in bins

```{r}

clust.plot <-  function(data, pal, title) {
  ggplot(data) +
    geom_bar(aes(fill = as.character(Year), y=pool.mean, x=Size),  position="dodge", stat="identity", alpha = 0.8) +
    geom_errorbar(aes(x=Size, ymin=pool.mean - pool.se, ymax = pool.mean + pool.se, group = as.character(Year)), 
                  position="dodge", size = 0.3) +
  theme_ipsum() +
  scale_fill_manual(values = pal, name="Year") +
  xlab("Size bin (TL)") + ylab("Pooled mean biomass") + ggtitle(title)
}


clust.plot2 <-  function(data, title) {
  ggplot(data) +
    geom_bar(aes(fill = as.character(Year), y=pool.mean, x=Size),  position="dodge", stat="identity", alpha = 0.8) +
    geom_errorbar(aes(x=Size, ymin=pool.mean - pool.se, ymax = pool.mean + pool.se, group = as.character(Year)), 
                  position="dodge", size = 0.3) +
  theme_ipsum() +
  scale_fill_viridis(direction = -1, discrete = T, name="Year") +
  xlab("Size bin (TL)") + ylab("Pooled mean biomass") + ggtitle(title)
}

temp <- sam5 %>% filter(TAXONNAME == "Naso lituratus") 
temp2 <- sam10 %>% filter(TAXONNAME == "Naso lituratus") 
grid.arrange(clust.plot2(temp, "Naso lituratus"), clust.plot2(temp2, "Naso lituratus"), ncol = 2)



```

# Split dataset by species, loop to create species directories for each region, and place clustered plots of each bin width in folders

```{r}

mar5.spp <- split(mar5, mar5$TAXONNAME)
mar10.spp <- split(mar10, mar10$TAXONNAME)
mar.size.dir <- "Figures/Size/Weighted/Territory/Time/Guam/"


for (i in 1:length(mar5.spp)) {
      folder <- paste0(mar.size.dir,names(mar5.spp[i]))
        if (!dir.exists(folder)) {dir.create(folder)}
      setwd(folder)
        p = clust.plot2(data = mar5.spp[[i]], title = paste(names(mar5.spp)[i], "5cm bins"))
        q = clust.plot2(data = mar10.spp[[i]], title = paste(names(mar10.spp)[i], "10cm bins"))
        ggsave(filename = paste("Clust Bars ", names(mar5.spp)[i], ".jpg", sep=""), width = 6, height = 8, arrangeGrob(q, p))
  setwd(proj.dir)
}

sam5.spp <- split(sam5, sam5$TAXONNAME)
sam10.spp <- split(sam10, sam10$TAXONNAME)
sam.size.dir <- "Figures/Size/Weighted/Territory/Time/Samoa/"


for (i in 1:length(sam5.spp)) {
      folder <- paste0(sam.size.dir,names(sam5.spp[i]))
        if (!dir.exists(folder)) {dir.create(folder)}
      setwd(folder)
        p = clust.plot2(data = sam5.spp[[i]], title = paste(names(sam5.spp)[i], "5cm bins"))
        q = clust.plot2(data = sam10.spp[[i]], title = paste(names(sam10.spp)[i], "10cm bins"))
        ggsave(filename = paste("Clust Bars ", names(sam5.spp)[i], ".jpg", sep=""), width = 6, height = 8, arrangeGrob(q, p))
  setwd(proj.dir)
}

## Stopped here 
## Add layered density plot with time as separate distributions? to same? 

```
