==> 02.2-Size-Obs.Rmd <==
---
title: "02.2-Size-Obs"
author: "Kurt Ingeman"
date: "5/27/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load Libraries

```{r}
library(here)
library(tidyverse)
library(ggplot2)
library(rlang)
library(hrbrthemes)
library(viridis)

```

# Clear Environment

```{r}
rm(list = ls())
```

# Read in data and format

```{r}

df <- readRDS("~/github/JCR-FMP/Data/Raw/Strata/priorityspp.size.meancount.site.strataN2.rds") %>% 
  filter(mean.count > 0) %>% 
  mutate(TL = as.numeric(Size_cmTL)) %>% 
  mutate_if(is.character, as.factor) %>% 
  mutate(index.count = 4* mean.count) %>% 
  select(-mean.count) %>% 
  uncount(index.count)

df$TAXONNAME <- as.factor(recode(df$TAXONNAME, "Chlorurus spilurus/sordidus" = "Chlorurus spilurus")) 
df$ISLAND <- droplevels(df$ISLAND)

df.TL.mar <- df %>% 
  filter(REGION == "MARIAN")  %>% 
  mutate(Year = as.factor(OBS_YEAR)) %>%
  group_by(TAXONNAME) %>% 
  summarise(mu_SPP = mean(TL, na.rm=TRUE), 
            max_spp = max(TL)) 

df.TL.guam <- df %>% 
  filter(ISLAND == "Guam")  %>% 
  mutate(Year = as.factor(OBS_YEAR)) %>%
  group_by(TAXONNAME) %>% 
  summarise(mu_SPP = mean(TL, na.rm=TRUE), 
            max_spp = max(TL)) 

rm(list = ls(pattern = ".TL"))


df <- df %>% 
  mutate(Year = as.factor(OBS_YEAR)) %>% 
  group_by(REGION, Year) %>% 
  mutate(mu_reg_year = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup() %>% 
  group_by(ISLAND, Year) %>% 
  mutate(mu_isl_year = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup() %>% 
  group_by(TAXONNAME, Year) %>% 
  mutate(mu_spp_year = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup() %>% 
  mutate(final.spp =
           case_when(TAXONNAME == "Acanthurus lineatus" ~ "final",
                     TAXONNAME == "Naso lituratus" ~ "final",
                     TAXONNAME == "Caranx melampygus" ~ "final", 
                     TAXONNAME == "Kyphosus cinerascens" ~ "final", 
                     TAXONNAME == "Cheilinus undulatus" ~ "final", 
                     TAXONNAME == "Lethrinus olivaceus" ~ "final", 
                     TAXONNAME == "Lutjanus fulvus" ~ "final", 
                     TAXONNAME == "Chlorurus frontalis" ~ "final", 
                     TAXONNAME == "Scarus schlegeli" ~ "final", 
                     TAXONNAME == "Epinephelus merra" ~ "final", 
                     TRUE ~ "initial"
                     ))


df.mar <- df %>% 
  filter(REGION == "MARIAN")  %>% 
  mutate(Pop_Unpop = case_when(ISLAND == "Guam" ~ "Populated", 
                               ISLAND == "Saipan" ~ "Populated", 
                               ISLAND == "Tinian" ~ "Populated", 
                               ISLAND == "Rota" ~ "Populated", 
                               TRUE ~ "Unpopulated")) %>% 
  mutate(ISLAND = factor(ISLAND, levels = c("Guam", "Saipan", "Tinian", "Rota", "Aguijan", "AGS", "Pagan", "Agrihan", "Asuncion", "Maug", "Farallon de Pajaros"))) %>% 
  group_by(Pop_Unpop, TAXONNAME) %>% 
  mutate(mu_pop_spp = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup()
  

df.sam <- df %>% 
  filter(REGION == "SAMOA") %>% 
    mutate(Pop_Grad = case_when(ISLAND == "Tutuila" ~ "High", 
                         ISLAND == "Tau" ~ "Mid", 
                         ISLAND == "Ofu & Olosega" ~ "Mid", 
                         TRUE ~ "Low")) %>% 
  mutate(ISLAND = factor(ISLAND, levels = c("Tutuila", "Tau", "Ofu & Olosega", "Swains", "Rose"))) %>%
  group_by(Pop_Grad, TAXONNAME) %>% 
  mutate(mu_pop_spp = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup()

df.sam$Pop_Grad <- factor(df.sam$Pop_Grad, levels = c("High", "Mid", "Low"))

```

# Region-level size structure; aggregate all species, facet by year to compare change through time

```{r}

size.obs.mar <- ggplot(data=df.mar, aes(x=TL, group=REGION, fill=Year)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_reg_year, color = Year),
             linetype = "dashed") +
      theme_ipsum() +
   scale_fill_viridis(discrete = TRUE, direction = -1) +
  scale_color_viridis(discrete = TRUE, direction = -1) +
  ggtitle("Mariana Islands, All Priority Species") +
  xlim(0,75) +
  facet_wrap(~Year, ncol = 1)

ggsave(size.obs.mar, filename = "Figures/Size/Obs/Territory/Mariana_year.jpg", width = 6, height = 6, dpi = 300, units = "in")


size.obs.sam <- ggplot(data=df.sam, aes(x=TL, group=REGION, fill=Year)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_reg_year, color = Year),
             linetype = "dashed") +
      theme_ipsum() +
   scale_fill_viridis(discrete = TRUE, direction = -1) +
  scale_color_viridis(discrete = TRUE, direction = -1) +
  ggtitle("Am Samoa, All Priority Species") +
  xlim(0,75) +
  facet_wrap(~Year, ncol = 1)

ggsave(size.obs.sam, filename = "Figures/Size/Obs/Territory/Samoa_year.jpg", width = 6, height = 8, dpi = 300, units = "in")

```

```{r}

rm(list = ls(pattern = "size"))

```

####
#### Species-level size structure; facet by species and aggregate across time and island
####

```{r}

df.mar <- df.mar %>% 
  group_by(TAXONNAME) %>% 
  mutate(mu_spp = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% 
  filter(!is.na(TAXONNAME)) %>% 
  mutate(TAXONNAME = fct_reorder(TAXONNAME, mu_spp))

df.mar.final <- df.mar %>% 
  filter(final.spp == "final")

df.guam <- df.mar %>% 
  filter(ISLAND == "Guam") %>% 
  group_by(TAXONNAME) %>% 
  mutate(mu_spp_guam = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% 
  filter(!is.na(TAXONNAME)) %>% 
  mutate(TAXONNAME = fct_reorder(TAXONNAME, mu_spp_guam))

df.guam.final <- df.guam %>% 
  filter(final.spp == "final")

df.sam <- df.sam %>% 
  group_by(TAXONNAME) %>% 
  mutate(mu_spp = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% 
  filter(!is.na(TAXONNAME)) %>% 
  mutate(TAXONNAME = fct_reorder(TAXONNAME, mu_spp))

p.size.obs.guam.spp <- ggplot(data=df.guam, aes(x=TL, group=TAXONNAME, fill=mu_spp_guam), show.legend = F) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_spp_guam),
             linetype = "dashed", show.legend = F) +
    scale_fill_viridis(direction = -1, ) +
  xlim(0,150) +
  theme_bw() +
   theme(legend.position="none") +
  facet_wrap(TAXONNAME ~ ., scales = "free")

ggsave(p.size.obs.guam.spp, filename = "Figures/Size/Obs/Species/Guam_spp.jpg", width = 12, height = 9, dpi = 300, units = "in")

##

p.size.obs.guam.final <- ggplot(data=df.guam.final, aes(x=TL, group=TAXONNAME, fill=mu_spp_guam), show.legend = F) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_spp_guam),
             linetype = "dashed", show.legend = F) +
    scale_fill_viridis(direction = -1, ) +
  xlim(0,150) +
  theme_bw() +
   theme(legend.position="none") +
  facet_wrap(TAXONNAME ~ ., scales = "free")

ggsave(p.size.obs.guam.final, filename = "Figures/Size/Obs/Species/Guam_spp_final.jpg", width = 12, height = 9, dpi = 300, units = "in")

##


p.size.obs.mar.spp <- ggplot(data=df.mar, aes(x=TL, group=TAXONNAME, fill=mu_spp), show.legend = F) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_spp),
             linetype = "dashed", show.legend = F) +
    scale_fill_viridis(direction = -1, ) +
  xlim(0,150) +
  theme_bw() +
   theme(legend.position="none") +
  facet_wrap(TAXONNAME ~ ., scales = "free")


ggsave(p.size.obs.mar.spp, filename = "Figures/Size/Obs/Species/Mariana_spp.jpg", width = 12, height = 9, dpi = 300, units = "in")

## Samoa
p.size.obs.sam.spp <- ggplot(data=df.sam, aes(x=TL, group=TAXONNAME, fill=mu_spp), show.legend = F) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_spp),
             linetype = "dashed", show.legend = F) +
    scale_fill_viridis(direction = -1, ) +
  xlim(0,110) +
  theme_bw() +
   theme(legend.position="none") +
  facet_wrap(TAXONNAME ~ ., scales = "free")


ggsave(p.size.obs.sam.spp, filename = "Figures/Size/Obs/Species/AmSamoa_spp.jpg", width = 12, height = 8, dpi = 300, units = "in")


```



####
#### Island-level size structure; facet by island and aggregate across time and species
####

```{r}

df.mar <- df.mar %>% 
  group_by(ISLAND) %>% 
  mutate(mu_isl = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% 
  filter(!is.na(ISLAND)) %>% 
  mutate(ISLAND = fct_reorder(ISLAND, mu_isl))

df.sam <- df.sam %>% 
  group_by(ISLAND) %>% 
  mutate(mu_isl = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% 
  filter(!is.na(ISLAND)) %>% 
  mutate(ISLAND = factor(ISLAND, levels = c("Tutuila", "Tau", "Ofu & Olosega", "Rose", "Swains")))

p.size.obs.mar.isl <-  ggplot(data=df.mar, aes(x=TL, group=ISLAND, fill=Pop_Unpop)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_isl),
             linetype = "dashed", show.legend = F) +
    scale_fill_viridis(discrete = TRUE, direction = 1, ) +
  guides(fill=guide_legend(title="Population level")) +
  ggtitle("Average size:\nPriority Species") +
  xlim(0,60) +
  theme_bw() +
  theme(legend.title=element_blank()) +
  facet_grid(ISLAND ~ .)


ggsave(p.size.obs.mar.isl, filename = "Figures/Size/Obs/Island/Mariana_isl.jpg", width = 8, height = 10, dpi = 300, units = "in")


p.size.obs.sam.isl <-  ggplot(data=df.sam, aes(x=TL, group=ISLAND, fill=Pop_Grad)) +
    geom_density(adjust=1.5, alpha=.4) +
  geom_vline(aes(xintercept = mu_isl, color = Pop_Grad),
             linetype = "dashed", show.legend = F) +
    scale_fill_viridis(discrete = TRUE, direction = 1, ) +
    scale_color_viridis(discrete = TRUE, direction = 1) +
  guides(fill=guide_legend(title="Population level")) +
  xlim(0,80) +
   theme_bw() +
  facet_grid(ISLAND ~ .)


ggsave(p.size.obs.sam.isl, filename = "Figures/Size/Obs/Island/Samoa_isl.jpg", width = 6, height = 8, dpi = 300, units = "in")


sam.pri <- df.sam %>% 
  mutate(final.pri = case_when(TAXONNAME == "Acanthurus lineatus" ~ "pri",
                               TAXONNAME == "Acanthurus xanthopterus" ~ "pri",
                               TAXONNAME == "Ctenochaetus striatus" ~ "pri",
                               TAXONNAME == "Myripristis berndti" ~ "pri",
                               TAXONNAME == "Monotaxis grandoculis" ~ "pri",
                               TRUE ~ "not")) %>% 
  filter(final.pri == "pri") %>% 
  group_by(ISLAND, TAXONNAME) %>% 
  summarise(mu_isl = mean(TL, na.rm=TRUE))

```





####
#### Region-level size structure; loop through each island, facet by year to compare change through time
####

# plot for loop split by island
```{r}

mar.obs.isl <- "Figures/Size/Obs/Island/Marianas/"
sam.obs.isl <- "Figures/Size/Obs/Island/AmSamoa/"

df.mar %>% 
  group_split(ISLAND) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=Year) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_isl_year, color = Year),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = first(.x$ISLAND)) +
    xlim(0,75) +
    facet_wrap(~Year, ncol = 1) +
    theme_ipsum() -> plot
    ggsave(paste0(mar.obs.isl, first(.x$ISLAND), '.jpg'), plot)
      })

df.sam %>% 
  group_split(ISLAND) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=Year) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_isl_year, color = Year),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = first(.x$ISLAND)) +
    xlim(0,75) +
    facet_wrap(~Year, ncol = 1) +
    theme_ipsum() -> plot
    ggsave(paste0(sam.obs.isl, first(.x$ISLAND), '.jpg'), plot)
      })




```


####
#### Species-level size structure; loop through each species, facet by year to compare change through time
####

# plot for loop split by species
```{r}

mar.obs.spp <- "Figures/Size/Obs/Species/Marianas/"
sam.obs.spp <- "Figures/Size/Obs/Species/AmSamoa/"

df.mar %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=Year) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_spp_year, color = Year),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = first(.x$TAXONNAME)) +
    xlim(0,75) +
    facet_wrap(~Year, ncol = 1) +
    theme_ipsum() -> plot
    ggsave(paste0(mar.obs.spp, first(.x$TAXONNAME), '.jpg'), plot)
      })

df.sam %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=Year) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_spp_year, color = Year),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = first(.x$TAXONNAME)) +
    xlim(0,75) +
    facet_wrap(~Year, ncol = 1) +
    theme_ipsum() -> plot
    ggsave(paste0(sam.obs.spp, first(.x$TAXONNAME), '.jpg'), plot)
      })

```


####
####  Size structure of each species by populated and unpopulated 
####


```{r message=FALSE}

mar.obs.pop.spp <- "Figures/Size/Obs/Pop/Marianas/"
sam.obs.pop.spp <- "Figures/Size/Obs/Pop/AmSamoa/"


df.mar %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=Pop_Unpop, group = Pop_Unpop) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_pop_spp, color = Pop_Unpop),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = 1) +
    scale_color_viridis(discrete = TRUE, direction = 1) +
    labs(x = "TL",
    y = "Frequency",
    title = paste(first(.x$TAXONNAME), "in Mariana Archipelago")) +
    theme_ipsum() -> plot
    ggsave(paste0(mar.obs.pop.spp, first(.x$TAXONNAME), '.jpg'), plot)
      })

df.sam %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=Pop_Grad, group = Pop_Grad) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_pop_spp, color = Pop_Grad),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = 1) +
    scale_color_viridis(discrete = TRUE, direction = 1) +
    labs(x = "TL",
    y = "Frequency",
    title = paste(first(.x$TAXONNAME), "in islands of Am Samoa")) +
    theme_ipsum() -> plot
    ggsave(paste0(sam.obs.pop.spp, first(.x$TAXONNAME), '.jpg'), plot)
      })

```

####
#### Species-level size structure; loop through each species, facet by depth ONLY to look for management relevant paterns 
####

```{r}

df.mar <- df.mar %>% 
  group_by(TAXONNAME, DEPTH_BIN) %>% 
  mutate(mu_depth_spp = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup()

df.mar$DEPTH_BIN <- factor(df.mar$DEPTH_BIN, levels = c("Shallow", "Mid", "Deep"))

df.sam <- df.sam %>% 
  group_by(TAXONNAME, DEPTH_BIN) %>% 
  mutate(mu_depth_spp = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup()

df.sam$DEPTH_BIN <- factor(df.sam$DEPTH_BIN, levels = c("Shallow", "Mid", "Deep"))

```

# plot for loop split by species

```{r}
mar.obs.depth.spp <- "Figures/Size/Obs/Depth/Marianas/"
sam.obs.depth.spp <- "Figures/Size/Obs/Depth/AmSamoa/"

df.mar %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=DEPTH_BIN, group = DEPTH_BIN) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_depth_spp, color = DEPTH_BIN),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = paste(first(.x$TAXONNAME), "in Mariana Archipelago")) +
    facet_wrap(~DEPTH_BIN, ncol = 1) +
    theme_ipsum() -> plot
    ggsave(paste0(mar.obs.depth.spp, first(.x$TAXONNAME), '.jpg'), plot)
      })

df.sam %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=DEPTH_BIN, group = DEPTH_BIN) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_depth_spp, color = DEPTH_BIN),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = paste(first(.x$TAXONNAME), "in islands of Am Samoa")) +
    facet_wrap(~DEPTH_BIN, ncol = 1) +
    theme_ipsum() -> plot
    ggsave(paste0(sam.obs.depth.spp, first(.x$TAXONNAME), '.jpg'), plot)
      })

```



####
#### Species-level size structure; loop through each species, facet by depth AND year to look for management relevant paterns 
####

```{r}

df.mar <- df.mar %>% 
  group_by(TAXONNAME, DEPTH_BIN, Year) %>% 
  mutate(mu_depth_spp_yr = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup() %>% ungroup()

df.mar$DEPTH_BIN <- factor(df.mar$DEPTH_BIN, levels = c("Shallow", "Mid", "Deep"))

df.sam <- df.sam %>% 
  group_by(TAXONNAME, DEPTH_BIN, Year) %>% 
  mutate(mu_depth_spp_yr = mean(TL, na.rm=TRUE)) %>%
  ungroup() %>% ungroup() %>% ungroup()

df.sam$DEPTH_BIN <- factor(df.sam$DEPTH_BIN, levels = c("Shallow", "Mid", "Deep"))

```

# plot for loop split by species

```{r}
mar.obs.depth.spp.yr <- "Figures/Size/Obs/Depth/Marianas/DepthByYear/"
sam.obs.depth.spp.yr <- "Figures/Size/Obs/Depth/AmSamoa/DepthByYear/"

df.mar %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=DEPTH_BIN, group = DEPTH_BIN) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_depth_spp_yr, color = DEPTH_BIN),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = paste(first(.x$TAXONNAME), "in Mariana Archipelago")) +
    facet_grid(DEPTH_BIN ~ Year) +
    theme_ipsum() -> plot
    ggsave(paste0(mar.obs.depth.spp.yr, first(.x$TAXONNAME), '.jpg'), plot)
      })

df.sam %>% 
  group_split(TAXONNAME) %>%
  map(~{ggplot(.x) + aes(x=TL, fill=DEPTH_BIN, group = DEPTH_BIN) +
    geom_density(adjust=1.5, alpha=.4) +
    geom_vline(aes(xintercept = mu_depth_spp_yr, color = DEPTH_BIN),
             linetype = "dashed") +
    scale_fill_viridis(discrete = TRUE, direction = -1) +
    scale_color_viridis(discrete = TRUE, direction = -1) +
    labs(x = "TL",
    y = "Frequency",
    title = paste(first(.x$TAXONNAME), "in islands of Am Samoa")) +
    facet_grid(DEPTH_BIN ~ Year) +
    theme_ipsum() -> plot
    ggsave(paste0(sam.obs.depth.spp.yr, first(.x$TAXONNAME), '.jpg'), plot)
      })

```
